% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DACT.R
\name{DACT}
\alias{DACT}
\title{DACT is a function used for testing for mediation effect in genome-wide epigenetic studies.}
\usage{
DACT(p_a,p_b,correction)
}
\arguments{
\item{p_a, }{the p-value vector for the exposure-mediator associations}

\item{p_b, }{the p-value vector for the mediator-outcome associations}

\item{correction, }{the name of method for correction, either "Efron" or "JC". By default, it is set to NULL.}
}
\value{
the p-value vector of mediation effect testing
}
\description{
DACT is a function used for testing for mediation effect in genome-wide epigenetic studies.
}
\references{
Liu, Z. , Shen, J., Barfield, R., Schwartz, J., Baccarelli, A., Lin, X., 2020. Large-scale hypothesis testing for causal mediation effects with applications in genome-wide epigenetic studies. (under review)
}
\author{
Zhonghua Liu < zhhliu@hku.hk >
}
